refactor(kernel): split plugin ir runtime seam
